Update sockssocket.py to newer version
Update to https://code.google.com/p/socksipy-branch/
This is fork of original project which was abandoned.

This update fixes the problem with interaction between srnd and i2pd.
i2pd socks5 behavior is different from java i2p. Exactly, i2pd forms
reply with DOMAINNAME address type (see
https://www.ietf.org/rfc/rfc1928.txt section "6. Replies" for details).
It puts a size of the address inside first byte of BND.ADDR.

This byte should be treated as binary and converted to int. In new
version of this module this task is accomplished by using ord().